+++ domain/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod message_sender;
+pub mod models;
 pub mod repository;
